[#145] refactor(chat): ChatRoomService에서 그룹/개인 채팅 책임을 분리하고 위임 구조로 정리
