+++ src/main/java/com/dduru/gildongmu/chat/service/PrivateChatRoomService.java
@@ -1,0 +1,83 @@
+package com.dduru.gildongmu.chat.service;
+
+import com.dduru.gildongmu.chat.domain.ChatRoom;
+import com.dduru.gildongmu.chat.domain.ChatRoomMember;
+import com.dduru.gildongmu.chat.domain.enums.ChatMemberRole;
+import com.dduru.gildongmu.chat.domain.enums.ChatRoomStatus;
+import com.dduru.gildongmu.chat.domain.enums.ChatRoomType;
+import com.dduru.gildongmu.chat.dto.response.PrivateChatRoomCreateResponse;
+import com.dduru.gildongmu.chat.exception.NotSelfChatException;
+import com.dduru.gildongmu.chat.repository.ChatRoomMemberRepository;
+import com.dduru.gildongmu.chat.repository.ChatRoomRepository;
+import com.dduru.gildongmu.post.domain.Post;
+import com.dduru.gildongmu.post.repository.PostRepository;
+import com.dduru.gildongmu.user.domain.User;
+import com.dduru.gildongmu.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+import java.util.Optional;
+
+@Slf4j
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class PrivateChatRoomService {
+
+    private final ChatRoomRepository chatRoomRepository;
+    private final ChatRoomMemberRepository chatRoomMemberRepository;
+    private final PostRepository postRepository;
+    private final UserRepository userRepository;
+
+    @Transactional
+    public PrivateChatRoomCreateResponse createOrGetPrivateRoom(Long requesterId, Long postId) {
+        Post post = postRepository.getActiveByIdOrThrow(postId);
+        User target = post.getUser();
+        User requester = userRepository.getByIdOrThrow(requesterId);
+        validateNotSelfPrivateChat(requesterId, target.getId());
+
+        Optional<ChatRoom> existingRoom = findExistingPrivateRoom(post, requester, target);
+        if (existingRoom.isPresent()) {
+            return new PrivateChatRoomCreateResponse(existingRoom.get().getId(), false);
+        }
+
+        Long newRoomId = createPrivateRoom(post, requester, target);
+        return new PrivateChatRoomCreateResponse(newRoomId, true);
+    }
+
+    private Optional<ChatRoom> findExistingPrivateRoom(Post post, User requester, User target) {
+        return chatRoomRepository.findPrivateRoomByPostAndUsers(
+                post.getId(),
+                ChatRoomType.PRIVATE,
+                ChatRoomStatus.ACTIVE,
+                requester,
+                target
+        );
+    }
+
+    private Long createPrivateRoom(Post post, User requester, User target) {
+        ChatRoom room = ChatRoom.forPrivateChat(post);
+        chatRoomRepository.save(room);
+        savePrivateRoomMembers(room, requester, target);
+        log.info("1:1 채팅방 생성 - roomId={}, postId={}, users=[{}, {}]", room.getId(), post.getId(),
+                requester.getId(), target.getId());
+        return room.getId();
+    }
+
+    private void savePrivateRoomMembers(ChatRoom room, User requester, User target) {
+        chatRoomMemberRepository.saveAll(List.of(
+                ChatRoomMember.create(room, requester, ChatMemberRole.HOST),
+                ChatRoomMember.create(room, target, ChatMemberRole.GUEST)
+        ));
+    }
+
+    private static void validateNotSelfPrivateChat(Long requesterId, Long targetUserId) {
+        if (requesterId.equals(targetUserId)) {
+            throw new NotSelfChatException();
+        }
+    }
+}
+
